Add the original ANSI Lisp tests to the files.
Did this after digging out the original files for "Did the original *really*
say that?" one too many times. I'd really like to be able to delete the original
when the perl version passes, but I'll let the SKIP markers handle that for me.

Also have a vague notion of how to write a different test suite style, but
that's yak shaving for later.